Rename and improve
