+++ src/main/integrations/sentry-minidump/index.ts
@@ -15,7 +15,7 @@ import { Store } from '../../store';
 import { getEventDefaults } from '../../context';
 import { checkPreviousSession, sessionCrashed } from '../../sessions';
 import { sentryCachePath } from '../../fs';
-import { getRendererState, trackRendererStates } from '../../renderer-state';
+import { getRendererProperties, trackRendererProperties } from '../../renderers';
 
 /** Sends minidumps via the Sentry uploader */
 export class SentryMinidump implements Integration {
@@ -58,7 +58,7 @@ export class SentryMinidump implements Integration {
       throw new SentryError('Attempted to enable Electron native crash reporter but no DSN was supplied');
     }
 
-    trackRendererStates();
+    trackRendererProperties();
 
     // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
     const transport = (client as any)._getBackend().getTransport() as ElectronNetTransport;
@@ -116,7 +116,7 @@ export class SentryMinidump implements Integration {
     const event = mergeEvents(await getEventDefaults(release), {
       contexts: {
         electron: {
-          crashed_url: getRendererState(contents.id)?.url || 'unknown',
+          crashed_url: getRendererProperties(contents.id)?.url || 'unknown',
           details,
         },
       },
